+++ pytorch_lightning/metrics/__init__.py
@@ -11,6 +11,8 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+from warnings import warn
+
 from pytorch_lightning.metrics.classification import (  # noqa: F401
     Accuracy,
     AUC,
@@ -38,3 +40,8 @@ from pytorch_lightning.metrics.regression import (  # noqa: F401
     SSIM,
 )
 from pytorch_lightning.metrics.retrieval import RetrievalMAP  # noqa: F401
+
+warn(
+    "`pytorch_lightning.metrics.*` module has been renamed to `torchmetrics.*` and split off to its own package"
+    " (https://github.com/PyTorchLightning/metrics) since v1.3 and will be removed in v1.5", DeprecationWarning
+)